meddra_rag_assistant/requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24
pandas>=2.1
pyyaml>=6.0
requests>=2.31
sentence-transformers>=2.6
streamlit>=1.31
openai>=1.6
python-dotenv>=1.0
rapidfuzz>=3.6
langchain>=0.2
chromadb>=0.4
torch>=2.0
langchain-community>=0.2
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
